feat: Adding support for running pattern tests over a range of LBAs
Adding support for write-read-compare over a range of LBAs.
A single flush cache is issued between the writes and reads to ensure everything comes from the medium.
I left code to disable write cache commented out as we need to decide if we want to do that in this function, let the caller disable it first, or if we want a flag to use FUA, etc.

With write cache disabled, this is very slow. With it enabled and a single flush it used, it is not too long.
Example times for walking 1's and 0's:
1. wcd = 45 mins
2. wce = 2.3 mins
3. wce + flush = 6 mins

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>